requirements.txt
Django==5.0.1
environs[django]==10.3.0
factory-boy==3.3.0
crispy-bootstrap5==2023.10
djangorestframework==3.14.0
django-oauth-toolkit==2.3.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9